also closes #282  and closes #284

## Changes

- Nuked lexer, as it was complicating things more then it was helping
- Rewritten parser, more graceful + should be easier to modify if needed
- Refactor PO struct, removed the fields that are not used anywhere. If needed can be added
- Added integration test, where specific commit superset PO file is loaded, and tested against some known inputs (headers ,and 3 messages)
- moved PO marshal to separate file

## Benchmark
|                    | Old implementation      | New Implementation                                      |
| ------------------ | ------------ | -------------------------------------------- |
|                    | **sec/op**       | **sec/op     vs base**                           |
| _Parse/Superset-10 | 2293.9µ ± 1% | 541.7µ ± 1%        -76.39% (p=0.000 n=10+20) |
| _Parse/Odoo-10     | 458.8µ ± 3%  | 191.5µ ± 1%        -58.25% (p=0.000 n=10+20) |
| geomean            | 1.026m       | 322.1µ              -68.60%                  |

|                    | Old implementation      | New Implementation                                      |
| ------------------ | ------------- | -------------------------------------------- |
|                    | B/op          | B/op      vs base                            |
| _Parse/Superset-10 | 9.530Mi ± 0%  | 6.105Mi ± 0%       -35.94% (p=0.000 n=10+20) |
| _Parse/Odoo-10     | 1103.6Ki ± 0% | 783.8Ki ± 0%       -28.98% (p=0.000 n=10+20) |
| geomean            | 3.205Mi       | 2.162Mi            -32.55%                   |

|                    | Old implementation     | New Implementation                              |
| ------------------ | ----------- | ------------------------------------ |
|                    | allocs/op   | allocs/op   vs base                  |
| _Parse/Superset-10 | 46.73k ± 0% | 11.79k ± 0%        -74.77% (n=10+20) |
| _Parse/Odoo-10     | 8.407k ± 0% | 2.281k ± 0%        -72.87% (n=10+20) |
| geomean            | 19.82k      | 5.186k             -73.84%           |