@janishorsts 2024-02-23
In what sense lexer was not working when lexing character by character?