@VladislavsPerkanuks 2024-02-26
Removing Regex yields minimum improvements in benchmarks, and results in more nested/complex code

Metric | with_regex.txt | without_tegex.txt | Comparison vs base
 -- | -- | -- | --
 sec/op | 1.727m ± 9% | 1.653m ± 3% | -4.25% (p=0.001 n=10)
 B/op | 3.934Mi ± 0% | 3.851Mi ± 0% | -2.11% (p=0.000 n=10)
allocs/op | 10.65k ± 0% | 10.56k ± 0% | -0.89% (p=0.000 n=10)